Remove Old Dev UI: Reactive messaging
