--- extensions/smallrye-reactive-messaging/deployment/src/main/resources/dev-templates/channels.html
@@ -1,44 +1,0 @@
-{#include main fluid=true}
-  {#style}
-   .annotation {
-     color: gray;
-   }
-  {/style}
-  {#title}Channels{/title}
-  {#body}
-  <table class="table table-striped ">
-   <thead class="thead-dark">
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">Name</th>
-      <th scope="col">Publisher</th>
-      <th scope="col">Consumers</th>
-    </tr>
-   </thead>
-   <tbody>
-    {#for channel in info:reactiveMessagingInfos.channels}
-    <tr>
-      <td>{channel_count}.</td>
-      <td><strong>{channel.name}</strong></td>
-      <td>{#componentDescription channel.publisher /}</td>
-      <td>
-        {#if channel.consumers}
-        {#when channel.consumers.size}
-          {#is 1}
-            {#componentDescription channel.consumers.get(0) /}
-          {#is > 1}
-            <ul>
-            {#for consumer in channel.consumers}
-              <li>{#componentDescription consumer /}</li> 
-            {/for}
-            </ul>
-          {#else}
-             N/A
-        {/when}
-        {/if}
-      </td>
-    {/for}
-   </tbody>
-  </table>
-  {/body}
-{/include}
