--- extensions/smallrye-reactive-messaging/deployment/src/main/resources/dev-templates/embedded.html
@@ -1,4 +1,0 @@
-<a href="{urlbase}/channels" class="badge badge-light">
-  <i class="fa fa-envelope"></i>
-  Channels <span class="badge badge-light">{info:reactiveMessagingInfos.channels.size}</span></a>
-<br>
