--- extensions/smallrye-reactive-messaging/deployment/src/main/resources/dev-templates/tags/componentDescription.html
@@ -1,17 +1,0 @@
-{#if it}
-{#when it.type}
-  {#case PUBLISHER}
-  <i class="fas fa-sign-out-alt"></i> Publisher:
-  {#case SUBSCRIBER}
-  <i class="fas fa-sign-in-alt"></i> Subscriber:
-  {#case PROCESSOR}
-  <i class="fas fa-exchange-alt"></i> Processor:
-  {#case CONNECTOR}
-  <i class="fas fa-plug"></i> Connector:
-  {#case EMITTER}
-  <i class="fas fa-syringe"></i> Emitter:
-  {#case CHANNEL}
-  <i class="fas fa-syringe"></i> Channel:
-{/when}
-{it.description.raw}
-{/if}
