+++ extensions/smallrye-reactive-messaging/deployment/src/main/resources/dev-ui/qwc-smallrye-reactive-messaging-channels.js
@@ -24,7 +24,7 @@ export class QwcSmallryeReactiveMessagingChannels extends LitElement {
     `;
 
     static properties = {
-        "_channels": {state: true, type: Array}
+        _channels: {state: true, type: Array}
     }
 
     /**
