--- extensions/smallrye-reactive-messaging/runtime/src/main/java/io/quarkus/smallrye/reactivemessaging/runtime/devconsole/DevReactiveMessagingInfosSupplier.java
@@ -1,12 +1,0 @@
-package io.quarkus.smallrye.reactivemessaging.runtime.devconsole;
-
-import java.util.function.Supplier;
-
-public class DevReactiveMessagingInfosSupplier implements Supplier<DevReactiveMessagingInfos> {
-
-    @Override
-    public DevReactiveMessagingInfos get() {
-        return new DevReactiveMessagingInfos();
-    }
-
-}
