+++ extensions/smallrye-reactive-messaging/runtime/src/main/java/io/quarkus/smallrye/reactivemessaging/runtime/devui/DevReactiveMessagingInfos.java
@@ -1,4 +1,4 @@
-package io.quarkus.smallrye.reactivemessaging.runtime.devconsole;
+package io.quarkus.smallrye.reactivemessaging.runtime.devui;
 
 import java.util.ArrayList;
 import java.util.Collections;
